tsplayer: bring up tsplayer for P. [1/1]
PD#SWPL-32550

Problem:
need porting tsplayer for P.

Solution:
porting code compatible with Q and P.

Verify:
ac214

SourceCode:
commit 24964db36a8b3cc95fbe41c5445e96ac10ea48cd
Author: shilong.yang <shilong.yang@amlogic.com>
Date:   Wed Sep 23 13:47:40 2020 +0800

+++ example/vesplayer/vesplayer.cpp
@@ -483,7 +483,7 @@ int main(int argc, char** argv) {
     }
 
     if (iname == nullptr ||
-        vFmt == -1) {
+        vFmt == (uint32_t)-1) {
         usage();
         exit(-1);
     }
